feat: Correção e finalização CRUD Pedidos

+++ GestaoPedidos/src/main/java/com/postech/orderservice/service/impl/PedidoServiceImpl.java
@@ -3,10 +3,12 @@ package com.postech.orderservice.service.impl;
 import com.postech.orderservice.entities.*;
 import com.postech.orderservice.repositories.ClienteRepository;
 import com.postech.orderservice.repositories.PedidoRepository;
+import com.postech.orderservice.repositories.ProdutoRepository;
 import com.postech.orderservice.service.PedidoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -14,11 +16,13 @@ public class PedidoServiceImpl implements PedidoService {
 
     private final PedidoRepository pedidoRepository;
     private final ClienteRepository clienteRepository;
+    private final ProdutoRepository produtoRepository;
 
     @Autowired
-    public PedidoServiceImpl(PedidoRepository pedidoRepository, ClienteRepository clienteRepository) {
+    public PedidoServiceImpl(PedidoRepository pedidoRepository, ClienteRepository clienteRepository, ProdutoRepository produtoRepository) {
         this.pedidoRepository = pedidoRepository;
         this.clienteRepository = clienteRepository;
+        this.produtoRepository = produtoRepository;
     }
 
     @Override
@@ -32,38 +36,54 @@ public class PedidoServiceImpl implements PedidoService {
     }
 
     @Override
-    public Pedido criarPedido(Pedido pedido) {
-        // Buscar o cliente no banco de dados
-        Cliente cliente = clienteRepository.findById(pedido.getCliente().getIdCliente())
-                .orElseThrow(() ->
-                        new IllegalArgumentException("Cliente não encontrado com ID: " + pedido.getCliente().getIdCliente()));
+    public Pedido criarPedido(Pedido pedido, Long idCliente) {
+        Cliente cliente = getClienteById(idCliente);
+        pedido.setCliente(cliente);
 
-        //BUSCAR PRODUTO
-
-        // Calcular o total do pedido
-
-        // Criar o pedido
-
-        // Salvar o pedido no banco de dados
-        return null; //pedidoRepository.save(pedido);
-
-    }
-
-    @Override
-    public Pedido atualizarPedido(Long id, Pedido pedidoNovo) {
-        Pedido pedido = this.buscarPedidoPorId(id);
-
-        pedido.setIdPedido(id);
-        pedido.setCliente(pedidoNovo.getCliente());
-        pedido.setItens(pedidoNovo.getItens());
-        pedido.setTotalPedido(pedidoNovo.getTotalPedido());
+        BigDecimal totalPedido = calcularTotalPedido(pedido.getItens());
+        pedido.setTotalPedido(totalPedido);
 
         return pedidoRepository.save(pedido);
     }
 
     @Override
+    public Pedido atualizarPedido(Long id, Pedido pedidoNovo) {
+        Pedido pedidoExistente = buscarPedidoPorId(id);
+
+        pedidoExistente.setCliente(pedidoNovo.getCliente());
+        pedidoExistente.setItens(pedidoNovo.getItens());
+        pedidoExistente.setTotalPedido(calcularTotalPedido(pedidoNovo.getItens()));
+
+        return pedidoRepository.save(pedidoExistente);
+    }
+
+    @Override
     public void deletarPedido(Long id) {
-        pedidoRepository.deleteById(id);
+        if (pedidoRepository.existsById(id)) {
+            pedidoRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Pedido não encontrado com ID: " + id);
+        }
+    }
+
+    private Cliente getClienteById(Long idCliente) {
+        return clienteRepository.findById(idCliente)
+                .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado com ID: " + idCliente));
+    }
+
+    private BigDecimal calcularTotalPedido(List<ItemPedido> itensPedido) {
+        BigDecimal total = BigDecimal.ZERO;
+        for (ItemPedido itemPedido : itensPedido) {
+            Produto produto = getProdutoById(itemPedido.getIdProduto());
+            BigDecimal subtotal = produto.getPreco().multiply(BigDecimal.valueOf(itemPedido.getQuantidade()));
+            total = total.add(subtotal);
+        }
+        return total;
+    }
+
+    private Produto getProdutoById(Long idProduto) {
+        return produtoRepository.findById(idProduto)
+                .orElseThrow(() -> new IllegalArgumentException("Produto não encontrado com ID: " + idProduto));
     }
 
 }
